Refactoring: move test classes to own package
https://github.com/metasfresh/metasfresh/issues/5611
